Replace `(PathBuf, bool)` with `CsvOptions`
I propose creating a new struct called `CsvOptions` that contains a set of options for saving data to a CSV file. Using `CsvOptions` will improve the readability of the code and better communicate the intention of the options being passed compared to using a tuple like `(PathBuf, bool)`. The struct contains the necessary fields to specify the path to the CSV file and whether timestamps should be formatted as absolute or relative times.