Add option to display multiple plots, and resize the plot panel
This adds the ability to display multiple plots, where each column can
be toggled individually per plot.

Additionally, it adds interactive resize using the seperator between the
plots and the serial monitor.

The resize code might not be the most pretty implementation, but I was
unable to find any example of this in the egui examples. Therefore this
commit also includes a small refactor of how heights are compted.

One thing left out of this commit, is the ability to save plots of any
but the first plot.