Fix theme preference not being restored correctly
Using `set_visuals(Visuals::dark())` in `main()` somehow interfered
badly with the theme switching code such that a light theme was not
restored properly.

Instead, the correct function to set the theme on startup is
`set_theme()` and we should pass in the saved perference instead of
hard-coding dark-mode.

With these changes, all of the theme preferences work correctly and are
restored correctly.